+++ src/api/index.js
@@ -15,16 +15,17 @@ const instance = axios.create({
   baseURL: BASE_URL
 });
 
-export const dispatchRequest = async (request) => {
+export const dispatchRequest = async (request, hideKeyboard = true) => {
   const _state = await state();
 
-  instance.defaults.headers.common['Content-Type'] = 'application/vnd.api+json';
-  instance.defaults.headers.common['Accept'] = 'application/vnd.api+json';
+  // instance.defaults.headers.common['Content-Type'] = 'application/vnd.api+json';
+  // instance.defaults.headers.common['Accept'] = 'application/vnd.api+json';
   // instance.defaults.headers.common['Authorization'] = _state.auth.token;
 
   try {
-    Keyboard.dismiss();
-
+    if (hideKeyboard) {
+      Keyboard.dismiss();
+    }
     const resource = await instance(request);
 
     return Promise.resolve(resource);