+++ src/components/Header/index.js
@@ -1,20 +1,23 @@
 import React, { Component } from 'react';
 import { View, Text, TouchableOpacity, SafeAreaView, TextInput } from 'react-native';
-import { Thumbnail, Label, Icon} from 'native-base';
+import { Thumbnail, Label, Icon } from 'native-base';
 import styles from './styles'
 const header = (props) => {
     return (
         <React.Fragment>
             {props.headerWithSearch && (
                 <SafeAreaView style={styles.container}>
-                    <View style={styles.searchFullContainer}>
+                    <View style={styles.fullHeaderContainer}>
+                        {/* <Image source={SearchIcon} style={styles.imgSearch} /> */}
                         <View style={styles.searchContainer}>
-                            <Image source={SearchIcon} style={styles.imgSearch} />
+                            <View style={styles.searchContainerIcon}>
+                                <Icon style={styles.serachIcon} type="FontAwesome"  name="search" />
+                            </View>
                             <TextInput
                                 editable={props.edit}
                                 style={styles.searchTextInput}
                                 placeholder={props.browseTitle}
-                                placeholderTextColor={colors.white}
+                                // placeholderTextColor={colors.white}
                                 onChangeText={value =>
                                     props.onChangeValue(props.stateInput, value)
                                 }