+++ src/components/SearchResult/styles.js
@@ -1,0 +1,13 @@
+import { StyleSheet, Platform } from 'react-native';
+import {
+    resize,
+} from '../../utils/styles';
+import fonts from '../../utils/fonts';
+import colors from '../../utils/colors';
+
+export default StyleSheet.create({
+    container: {
+        alignItems: 'center',
+        backgroundColor: colors.backgroundColorApp,
+    },
+});